Added character sheet to the left og the inventory Next we need to work on adding the search room --> function

+++ C_Programs/rpg/Headers/entityStructs.h
@@ -28,7 +28,7 @@ typedef struct Player{
     char *className;
 
     int level;
-    int gold;
+    double gold;
 
     Inventory *inventory;
